Move tests to grakn-test. Add docker-tests property.
This is so the reports get put in the correct directories and the tests
get run at the correct stage, instead of in grakn-test-profiles.
Property inheritance doesn't work properly like the surefire tests
because the bash script finds the tests manually and the directory is
hardcoded so tests always run.